requirements.txt
flask
opencv-python
numpy
ultralytics
paddleocr
mysql-connector-python
